pyproject.toml
[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[project]
name = "ingestor"
version = "0.1.0"
description = "Production-grade ingestion toolkit for LLM security datasets"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Apache-2.0" }
authors = [
  { name = "Raihan K. <raaihank@users.noreply.github.com>" }
]
dependencies = [
  "typer[all]>=0.9,<0.10",
  "click>=8.1,<8.2",
  "pydantic>=2,<3",
  "PyYAML>=6.0",
  "structlog>=24.1",
  "tenacity>=8.2",
  "rich>=13.7",
  "datasets>=2.19",
  "huggingface_hub>=0.23",
  "GitPython>=3.1",
  "kaggle>=1.6",
  "datasketch>=1.6",
  "langdetect>=1.0.9",
  "Unidecode>=1.3",
  "tqdm>=4.66",
  "fasttext-wheel>=0.9.0",
  "pandas>=2.0",
  "pyarrow>=15.0",
  "orjson>=3.10"

]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "mypy>=1.11",
  "ruff>=0.6",
  "types-PyYAML>=6.0.12.20240917",
  "build>=1.2",
]

[project.scripts]
ingestor = "ingestor.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["ingestor"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

